Move product models to single module
