--- saleor/product/models/__init__.py
@@ -1,7 +1,0 @@
-from .base import (AttributeChoiceValue, Category, Product, ProductAttribute,
-                   ProductVariant, Stock, ProductClass, StockLocation)
-from .images import ProductImage, VariantImage
-
-__all__ = ['AttributeChoiceValue', 'Category', 'Product', 'ProductAttribute',
-           'ProductImage', 'ProductVariant', 'Stock', 'VariantImage',
-           'ProductClass', 'StockLocation']
